Enable authorization for both versions of the api while deleting post at a router level using passport and enable authorisation at action/controller level aswell

+++ controllers/api/v1/posts_api.js
@@ -28,6 +28,8 @@ module.exports.destroy = async function(req,res){
     
 
     try {
+
+        
                 //first find the post exits or not depending on the url
         let post = await Post.findById(req.params.id);
         //this param will contai the id of the user who made the post
@@ -36,13 +38,14 @@ module.exports.destroy = async function(req,res){
         //user.id just covers our user._id into aa string automatically
         //now we check if the person deleting the post must be authour of the post 
         //we havent populated user yet , so it must be the user id
-      
+        
+        if(post.user == req.user.id ){
 
             //deleted the post 
             post.remove();
             //now go to the db and dlete aall the comments from that post 
             //the id of the post is still there in the param
-            post.update();
+          
             
             await Comments.deleteMany({post:req.params.id})
 
@@ -51,7 +54,11 @@ module.exports.destroy = async function(req,res){
           return res.json(200 , {
               message:"Post and associated comments delted !"
           })
-
+        }else{
+            return res.json(401 , {
+                message :"You cannot delete this post"
+            })
+        }
     
     }catch (err) {
